Special case cran metarepo

+++ R/meta.R
@@ -36,6 +36,7 @@ check_and_trigger <- function(universe){
 }
 
 needs_update <- function(universe){
+  if(cran) return("everything")
   retry(git_clone(paste0('https://github.com/r-universe/', universe)))
   fullpath <- normalizePath(universe)
   on.exit(unlink(fullpath, recursive = TRUE))
